requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
Flask==1.1.2
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
redis==3.4.1
requests==2.23.0
urllib3==1.25.8
Werkzeug==1.0.1
pyOpenSSL==23.2.0
python-jose==3.3.0
six~=1.16.0
dill==0.3.7
Authlib==1.2.1